serve: 📻 outline service creation from `Serve::exec()`
we will want this in code outside of the `Serve` command. so, move it
into the sender submodule.
