+++ src/opt/serve.rs
@@ -1,38 +1,15 @@
 use anyhow::Context;
 use clap::Parser;
-use directories::ProjectDirs;
 use futures::stream::FuturesUnordered;
-use futures_util::{stream::StreamExt, stream::TryStreamExt};
+use futures_util::stream::StreamExt;
 use num_traits::identities::Zero;
 use penumbra_asset::Value;
-use penumbra_custody::{soft_kms::SoftKms, CustodyClient};
-use penumbra_keys::FullViewingKey;
-use penumbra_proto::{
-    custody::v1::{
-        custody_service_client::CustodyServiceClient, custody_service_server::CustodyServiceServer,
-    },
-    view::v1::{view_service_client::ViewServiceClient, view_service_server::ViewServiceServer},
-};
-use penumbra_view::{ViewClient, ViewServer};
 use serenity::{http::Http, prelude::GatewayIntents};
-use std::{
-    env,
-    path::{Path, PathBuf},
-    sync::Arc,
-    time::Duration,
-};
+use std::{env, path::PathBuf, sync::Arc, time::Duration};
 use tokio::sync::mpsc;
-use tower::{
-    balance::p2c::Balance, limit::concurrency::ConcurrencyLimit, load::PendingRequestsDiscover,
-};
 use url::Url;
 
-use crate::{
-    opt::ChannelIdAndMessageId,
-    responder::RequestQueue,
-    sender::{SenderRequest, SenderSet},
-    Catchup, Handler, Responder, Sender, Wallet,
-};
+use crate::{opt::ChannelIdAndMessageId, responder::RequestQueue, Catchup, Handler, Responder};
 
 #[derive(Debug, Clone, Parser)]
 pub struct Serve {
@@ -81,7 +58,9 @@ impl Serve {
         let discord_token = env::var("DISCORD_TOKEN")?;
 
         // Make a worker to handle the address queue
-        let service = self.senders_service().await?;
+        let service =
+            crate::sender::service::init(self.data_dir.as_deref(), self.account_count, &self.node)
+                .await?;
         let (send_requests, responder) =
             Responder::new(service, self.max_addresses, self.values.clone());
         let (cancel_tx, mut cancel_rx) = mpsc::channel(1);
@@ -110,140 +89,6 @@ impl Serve {
             result = catch_up => result.context("error in catchup service")?,
             _ = cancel_rx.recv() => Err(anyhow::anyhow!("cancellation received")),
         }
-    }
-
-    /// Constructs a [`tower::Service`] of [`Sender`]s mapping requests to transaction hashes.
-    /* TODO(kate): the responder constructor hard-codes this return type. fix it later. */
-    async fn senders_service(
-        &self,
-    ) -> anyhow::Result<
-        ConcurrencyLimit<
-            Balance<
-                PendingRequestsDiscover<
-                    SenderSet<
-                        ConcurrencyLimit<
-                            Sender<
-                                impl ViewClient + Clone + 'static,
-                                impl CustodyClient + Clone + 'static,
-                            >,
-                        >,
-                    >,
-                >,
-                SenderRequest,
-            >,
-        >,
-    > {
-        // Create a sender set, with one sender for each account.
-        let d = self.senders_iter().await.map(SenderSet::new)?;
-
-        // Distribute requests across the senders.
-        let lb = Balance::new(PendingRequestsDiscover::new(
-            d,
-            tower::load::CompleteOnResponse::default(),
-        ));
-
-        // Limit the concurrent requests to 1-per-account.
-        let account_count = self
-            .account_count
-            .try_into()
-            .expect("number of accouts < u32::MAX");
-        let service = ConcurrencyLimit::new(lb, account_count);
-
-        Ok(service)
-    }
-
-    /// Returns an [`Iterator`] that will emit a [`Sender`] for each account index.
-    ///
-    /// This initializes a custody server and a view server that the senders will use to create and
-    /// authorize spends.
-    async fn senders_iter(
-        &self,
-    ) -> anyhow::Result<
-        impl Iterator<
-            Item = (
-                u32, // the account id
-                ConcurrencyLimit<
-                    Sender<
-                        // a rate-limited sender
-                        impl ViewClient + Clone + 'static,
-                        impl CustodyClient + Clone + 'static,
-                    >,
-                >,
-            ),
-        >,
-    > {
-        // Configure custody for Penumbra wallet key material.
-        // Look up the path to the view state file per platform, creating the directory if needed
-        let data_dir = self.data_dir.clone().unwrap_or_else(|| {
-            ProjectDirs::from("zone", "penumbra", "pcli")
-                .expect("can access penumbra project dir")
-                .data_dir()
-                .to_owned()
-        });
-        std::fs::create_dir_all(&data_dir).context("can create data dir")?;
-        tracing::debug!(?data_dir, "loading custody key material");
-
-        // Initialize a custody service client and a view service client for the senders.
-        let (custody, fvk) = Self::initialize_custody_service(&data_dir)?;
-        let view = Self::initialize_view_service(&self.node, &data_dir, &fvk).await?;
-
-        let make_sender = move |id| {
-            let sender = Sender::new(id, fvk.clone(), view.clone(), custody.clone());
-            (id, sender)
-        };
-
-        // Instantiate a sender for each account index.
-        Ok((0..self.account_count).map(make_sender))
-    }
-
-    /// Initializes a custody client, to request spend authorization.
-    fn initialize_custody_service(
-        data_dir: &Path,
-    ) -> anyhow::Result<(impl CustodyClient + Clone + Send + 'static, FullViewingKey)> {
-        let pcli_config_file = data_dir.join("config.toml");
-        let wallet = Wallet::load(pcli_config_file)
-            .context("failed to load wallet from local custody file")?;
-        let soft_kms = SoftKms::new(wallet.spend_key.clone().into());
-        let custody = CustodyServiceClient::new(CustodyServiceServer::new(soft_kms));
-        let fvk = wallet.spend_key.full_viewing_key().clone();
-        Ok((custody, fvk))
-    }
-
-    // Initializes a view client, to scan the Penumbra chain.
-    //
-    // This will build and synchronize the view service before it returns the client.
-    async fn initialize_view_service(
-        node: &Url,
-        data_dir: &Path,
-        fvk: &FullViewingKey,
-    ) -> anyhow::Result<impl ViewClient + Clone + Send + 'static> {
-        let view_file = data_dir.join("pcli-view.sqlite");
-        tracing::debug!("configuring ViewService against node {}", &node);
-        let view_filepath = Some(
-            view_file
-                .to_str()
-                .ok_or_else(|| anyhow::anyhow!("Non-UTF8 view path"))?
-                .to_string(),
-        );
-        let view_storage =
-            penumbra_view::Storage::load_or_initialize(view_filepath, &fvk, node.clone()).await?;
-        let view_service = ViewServer::new(view_storage, node.clone()).await?;
-
-        // Now build the view and custody clients, doing gRPC with ourselves
-        let mut view = ViewServiceClient::new(ViewServiceServer::new(view_service));
-
-        // Wait to synchronize the chain before doing anything else.
-        tracing::info!(
-            "starting initial sync: please wait for sync to complete before requesting tokens"
-        );
-        ViewClient::status_stream(&mut view)
-            .await?
-            .try_collect::<Vec<_>>()
-            .await?;
-        // From this point on, the view service is synchronized.
-        tracing::info!("initial sync complete");
-
-        Ok(view)
     }
 
     /// Configures a new discord [`Client`][serenity::Client].
